+++ src/webapp/context/ClusterMapContext.js
